+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased
+# v7.2.0 - 2022-11-25
 
 Changed
   * AZ-908: Use the new data source for CAF naming (instead of resource)